Don't serve RRDP files.
