+++ src/test/scala/net/ripe/rpki/publicationserver/MonitoringServiceTest.scala
@@ -3,14 +3,14 @@ package net.ripe.rpki.publicationserver
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import org.mockito.Mockito.{RETURNS_SMART_NULLS, when}
 
-class RrdpServiceTest extends PublicationServerBaseTest with ScalatestRouteTest with RRDPService  {
+class MonitoringServiceTest extends PublicationServerBaseTest with ScalatestRouteTest with MonitoringService  {
 
   trait Context {
     def actorRefFactory = system
   }
 
   test("should return a health check response") {
-    Get("/monitoring/healthcheck") ~> rrdpAndMonitoringRoutes ~> check {
+    Get("/monitoring/healthcheck") ~> monitoringRoutes ~> check {
       val response = responseAs[String]
       response should include ("commit")
     }
